Move GraphIndex class to root of routing package
