+++ src/test/java/org/opentripplanner/routing/core/StateEditorTest.java
@@ -2,7 +2,7 @@ package org.opentripplanner.routing.core;
 
 import org.junit.Test;
 import org.opentripplanner.routing.graph.Graph;
-import org.opentripplanner.routing.graph.GraphIndex;
+import org.opentripplanner.routing.GraphIndex;
 import org.opentripplanner.routing.impl.StreetVertexIndex;
 
 import static org.junit.Assert.assertEquals;
